pyproject.toml
[project]
name = "virtulingo-backend"
version = "0.1.0"
description = "Backend for Virtulingo, a language learning platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg==0.29.0",
    "deepgram-sdk==3.7.0",
    "fastapi==0.115.0",
    "faster-whisper==1.2.1",
    "google-genai>=1.2.0",
    "httpx==0.27.0",
    "openai==1.44.0",
    "pydantic>=2.8.2",
    "python-dotenv==1.0.1",
    "redis==5.0.8",
    "soundfile==0.13.1",
    "supabase>=2.28.0",
    "tts==0.22.0",
    "uvicorn[standard]==0.30.6",
]
